allow NOCLEANUP switch in all test-scripts for easier diagnostics and testing Original: 9e589f0ac34d21ab2853476fb95743e769dfce6c

+++ lalapps/src/pulsar/FDS_isolated/testCFSv2.sh
@@ -187,5 +187,6 @@ else
 fi
 
 ## clean up files
-rm -rf $SFTdir $outfile_v1 $outfile_v2NWon $outfile_v2NWoff Fstats Fstats.log
-
+if [ -z "$NOCLEANUP" ]; then
+    rm -rf $SFTdir $outfile_v1 $outfile_v2NWon $outfile_v2NWoff Fstats Fstats.log
+fi
