+++ lalapps/src/pulsar/FDS_isolated/testGridv2.sh
@@ -256,5 +256,6 @@ fi
 echo
 
 ## clean up files
-rm -rf $SFTdir $outputv2_0 $outputv2_1 $outputv2_2 $outputv1_0 $outputv1_1 $outputv1_2 Fstats Fstats.log
-
+if [ -z "$NOCLEANUP" ]; then
+    rm -rf $SFTdir $outputv2_0 $outputv2_1 $outputv2_2 $outputv1_0 $outputv1_1 $outputv1_2 Fstats Fstats.log
+fi
